Refactor element attributes retrieval
